ASRC-223: move VistaDaoFactory to the security package
because it is only used there and it clarifies its purpose.
