+++ srcalc/src/main/java/gov/va/med/srcalc/security/VistaLinkVistaDaoFactory.java
@@ -1,4 +1,7 @@
-package gov.va.med.srcalc.vista;
+package gov.va.med.srcalc.security;
+
+import gov.va.med.srcalc.vista.VistaDao;
+import gov.va.med.srcalc.vista.VistaLinkVistaDao;
 
 public class VistaLinkVistaDaoFactory implements VistaDaoFactory
 {
